now compiles under amd64

+++ trunk/source/engine/cl_cin.c
@@ -33,6 +33,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 #include "client.h"
 #include "snd_local.h"
+#include <stdint.h>
 
 #define MAXSIZE				8
 #define MINSIZE				4
@@ -1232,8 +1233,8 @@ static void readQuadInfo( byte *qData )
     cinTable[currentHandle].VQ0 = cinTable[currentHandle].VQNormal;
     cinTable[currentHandle].VQ1 = cinTable[currentHandle].VQBuffer;
 
-    cinTable[currentHandle].t[0] = (0 - (unsigned int)cin.linbuf)+(unsigned int)cin.linbuf+cinTable[currentHandle].screenDelta;
-    cinTable[currentHandle].t[1] = (0 - ((unsigned int)cin.linbuf + cinTable[currentHandle].screenDelta))+(unsigned int)cin.linbuf;
+    cinTable[currentHandle].t[0] = (0 - (intptr_t)cin.linbuf)+(intptr_t)cin.linbuf+cinTable[currentHandle].screenDelta;
+    cinTable[currentHandle].t[1] = (0 - ((intptr_t)cin.linbuf + cinTable[currentHandle].screenDelta))+(intptr_t)cin.linbuf;
 
     cinTable[currentHandle].drawX = cinTable[currentHandle].CIN_WIDTH;
     cinTable[currentHandle].drawY = cinTable[currentHandle].CIN_HEIGHT;
